requirements.txt
appdirs==1.4.3
attrs==19.2.0
beautifulsoup4==4.8.0
black==19.3b0
Click==7.0
Django==2.2.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-jquery==3.1.0
django-mathfilters==0.4.0
django-multiselectfield==0.1.9
pytz==2019.2
soupsieve==1.9.4
sqlparse==0.3.0
toml==0.10.0
django-extensions